Memory checker with hashmap

+++ include/sbi/sbi_step_ht.h
@@ -6,7 +6,7 @@
 // redefine these values for custom types
 
 typedef uint64_t ht_key_t;
-typedef uint8_t ht_val_t;
+typedef uint64_t ht_val_t;
 
 static inline uint64_t ht_hash(ht_key_t key) {
     key = ((key >> 16) ^ key) * 0x119de1f3;
